Fetch Olin phone charger data from libservices
All of the caveats from the earlier commit for laptops (42d0f7e) applies
here as well.

Since the equipment feed is even messier than laptops, more massaging is
required, which has been handed off to a new libservices util in an
attempt to keep the store mutations and actions clean.