First pass at oku-circ component
Call it twice from the page, passing the location prop and this results
in rendering the status for both locations.

This includes updating the libservices util to support dynamic API calls
(both in terms of endpoints and location ID).

Couldn't set the API_URL env variable as eluded to in the nuxt-helpers
doc, so resorted to declaring a constant within the util. This is all
just proof-of-concept for now anyway...everything is subject to change.

https://github.com/fandogh/nuxt-helpers#-usage

It should be noted that the location passed in the dynamic route
`/oku/:location` has no consequence on the oku-circ component just yet;
Olin will always be rendered first. Similarily, the page's fetch method
will always dispatch the laptops and phoneChargers actions for both Olin
and Uris, as it is also no longer `location` aware. This is a temporary
hack, not a feature ;)