First pass at Mann consultation screens
Bare minimum. Hit LibCal API and store response in Vuex for single desk.
Temporarily using API's `currently_open` field to determine status,
which we know is a big tease [1]...if only it were that simple.

BG color change based on status via Nuxt's header method, which stands
on the shoulders of `vue-meta`. Use fiord (`#485963`) from original
mockups for `by-appointment` status, which we hadn't considered in early
phases, but GIS holds this status the entire summer. No styling beyond
this and Lato font.

[1] cul-it/circ-display@53e6f4e