Test if the desk's change in status has past
Necessary now that we're caching the LibCal Hours API for two
minutes...this will force an update if we cross the `statusChange`
threshold even if the cache has not yet expired.

Now storing statusChange as ISO timestamp instead of formatted string.
Formatting is handled by dedicated method, called by page component.
Side effect is that the libcal util needs to be imported in the page
component (figured this was the lesser of two evils -- vs. importing
Moment.js, which cannot be cherry-picked), but went with this for now
over storing two varieties of statusChange.

Will revisit down the road to determine best approach...