Cache LibCal hours API response for 2 minutes  (#41)
* Update desk status every 30s

Since LibCal doesn't set Access-Control-Allow-Origin header, CORS is not
supported. So when fetching updates from client, need to deal with
JSONP. Luckily jsonp-promise to the rescue.

Also includes a change to send a payload to the Vuex action/dispatch so
multiple params can be passed.

* Need to pass jsonp bool to openNow() as well

Follow up to previous commit (d05290b). Since getHours() is eventually
called via openNow() -> nextOpening() -> openingTime() and once again we
need to resort to JSONP when hitting LibCal API from the client.

* Cache LibCal Hours API response for 2 mins #38

Use Vuex store as the cache. The `fetchStatus` action is still
dispatched every 30 seconds. TODO to convert staleCache() to
async/await.

* Test if the desk's change in status has past

Necessary now that we're caching the LibCal Hours API for two
minutes...this will force an update if we cross the `statusChange`
threshold even if the cache has not yet expired.

Now storing statusChange as ISO timestamp instead of formatted string.
Formatting is handled by dedicated method, called by page component.
Side effect is that the libcal util needs to be imported in the page
component (figured this was the lesser of two evils -- vs. importing
Moment.js, which cannot be cherry-picked), but went with this for now
over storing two varieties of statusChange.

Will revisit down the road to determine best approach...

* Test if last status update was yesterday

If so, force status refresh to ensure relative openings are correct.

* Bugfix: capture minutes in format string

While working on caching, noticed that next opening time for Knight was
incorrect (3:00pm instead of 3:30pm). This simplifies code down to
single format string for moment, as it should have been from the get go.

* Remove debugging

Cleanup on aisle 6.