Iterate bookings, add available slots btw & after
First pass at this. Not accounting for leading available slots since
we're still working with mock data and unsure whether the API
drops/excludes bookings once current time is greater than the `toDate`.

Early testing indicates this is not the case (any booking with a
datetime for the requested date remains), which will mean more logic
needs to be introduced to handle open and current availability.