Extract LibCal schema & generalize hour functions
First step in cleaning house and modularizing LibCal utils. Potential to
eventually extract into separate package.

Also generalize the hours related functions to use `location` in place
of 'desk'. Still a bit awkward with passing of the `isDesk` param to
handle different lookup in the schema, but wanted to keep the
consultation desks grouped together. Most likely will revisit this down
the road.