Remove whileOpen filter (spaces schedule)
This is a safeguard that was necessary in the legacy system but we may
be able to live without in the new app. Time will tell, but realized
this was preventing multiple-day reservations from appearing (which
technically shouldn't happen) since our policies disallow it, but case
in point with the interview room moving and Tom reserved it for 1 mo+.

Also, this filter was going to need to be disabled selectively for the
B30 classrooms at the very least since we have a longstanding policy of
allowing reservations in these classrooms when the basement is actually
closed. [1]

[1] cul-it/circ-display@ff893b4