Reverse sort spaces in computed property
As opposed to attempting to sort as part of the mutation. Not a good
idea when dealing with numeric keys for the store's object, which will
always result in sorting incrementally.

Noticed this bug while working on B30 display and realizing that reverse
was working for B30 classrooms (non numeric key thanks to the leading
"b"), but not working as originally implemented for Mann study rooms.