Add startedToday filter for LibCal bookings
Fixes bug introduced when removing whileOpen filter in #101.

The unexpected side-effect of that change was that bookings which
started the previous day but ran into the current day (including a
booking that ends at Midnight) were no longer being filtered out so the
available since opening test was failing due to an index shift (from 0
to 1). This resulted in an incorrect opening time of 12am being
displayed for any space where there was booking from the previous day
which ran until midnight). Clear as mud...

Don't fret, this is only a temporary bandage until we refactor for early
morning closings in #89.