Simplify logic for building/parsing LibCal & R25 space schedules (#108)
* Add name key for spaces store

Redundant to parent key but a common pattern and should be useful in
simplifying logic for updating store with parsed schedule/availability.

* Simplify logic for parsing/building space schedule

* Test for reservations before building a schedule

No need to parse/build if filtering leaves us with no eligible bookings.

* Add startedToday filter for LibCal bookings

Fixes bug introduced when removing whileOpen filter in #101.

The unexpected side-effect of that change was that bookings which
started the previous day but ran into the current day (including a
booking that ends at Midnight) were no longer being filtered out so the
available since opening test was failing due to an index shift (from 0
to 1). This resulted in an incorrect opening time of 12am being
displayed for any space where there was booking from the previous day
which ran until midnight). Clear as mud...

Don't fret, this is only a temporary bandage until we refactor for early
morning closings in #89.

* More versatile update mutation for spaces store

Utilize `Object.assign()` to add any/all new keys to existing store
instead of explicitly setting the schedule.

* Reduce redundancy btw variable & key

Department of redundancy reducing department...

* Apply simplified logic to r25 building/parsing

Still need to revisit these pieces and pull out the rest of the reusable
bits into a shared core but this is a good start.

* R25 spaces sans reservations are now covered

Thanks to refactoring from previous commit (d2996d9).

* TODO for pulling out common core