Adjust qualification for whileOpen filter
More leniency for this filter to accommodate bookings that span across
midnight. Instead of requiring that the booking falls completely within
the open and closing times for the space, now require that either the
start or end is between the open and closing time.

This is critical for supporting bookings that span midnight and goes
hand-in-hand with previous commit (b902bce).

Includes a comment for REVIEW:

Alternative to this approach -- Should we adjust the opening time (shift
it back to previous day opening) as we do for the closing?