Check if still open from yesterday (#111)
* Check if still open from yesterday #89

First pass...needs additional testing.

* Revisit startedToday filter after further testing

Update note to self.

* LibCal hours API limits past dates to current week

Instead of an error, API returns an empty array when requesting hours
for a date in the past beyond the current week (LibCal weeks start on
Sunday).

TODO: Reach out to Springshare support and request they increase this
limit to past 2 weeks

* Consistency for r25 import (conventions)

* Check tomorrow's bookings if LibCal space

Capture those early morning bookings that actually take place tomorrow
but prior to today's closing.

* Fetch tomorrow's bookings when applicable

Both conditions must pass:

* LibCal space

* Early morning closing (current boundaries: after midnight, before 6am)

Pull out test for early morning closing into separate function so it can
be utilized in the spaces store. Add TODO note about using this function
in the modified realClosingTime() as well.

* Return of the whileOpen filter

It's neccessary now that we're fetching reservations for following day
due to early morning closings. Still some open questions based on what
this filter suppresses:

* Should we care about multi-day reservations (not currently allowed)?

* What about our longstanding B30 policy of booking when the basement is
  closed and still wanting to display the event/reservation on a sign?

> Originally removed in #101

* Dedupe bookings that span over midnight

Necessary now that we fetch bookings for the following day for any space
that has an early morning closing.

* Use yesterday's closing time when applicable

If still open from yesterday, then use yesterday's closing time since it
hasn't happened yet. This is critical for supporting bookings that span
midnight.

* Adjust qualification for whileOpen filter

More leniency for this filter to accommodate bookings that span across
midnight. Instead of requiring that the booking falls completely within
the open and closing times for the space, now require that either the
start or end is between the open and closing time.

This is critical for supporting bookings that span midnight and goes
hand-in-hand with previous commit (b902bce).

Includes a comment for REVIEW:

Alternative to this approach -- Should we adjust the opening time (shift
it back to previous day opening) as we do for the closing?