Add source-map-loader as dependency
To address the 404 on mobile Safari 9 (iOS 9) for /.nuxt/es6-promise.map

This is a source map from nuxt dependency of es6-promise [1]. The 404 did
not appear to have any adverse effect of consequence on the app, so not
convinced this change is necessary. But figured it was worth cleaning up
the console while debugging for `const` keyword error addressed in
previous commit (4bb1ad0).

[1] https://github.com/stefanpenner/es6-promise/issues/235#issuecomment-406761142