Cocktail lounge edge case: opening with no closing
Yep...special. Cocktail Lounge is set to debut today and hours in libcal
are as follows:

{
  from: '8am',
  to: ''
}

Since the change in status is not determinable in this situtation,
update logic in SpaceAvailabilityItem component to simply not display
a status change under this circumstance.

Starting tomorrow (2019-06-11) cocktail lounge hours are 24/7 (with some
exceptions)...which means we'll have more edge cases to test later this
week :)