Init occupancy as data then update via fetch
Trying this alternative approach since the LiveOccupancy component
wasn't reacting to current Vuex state for occupancy when switching from
closed to open.

Still not 100% convinced this is a home run (I noticed a flicker of Max
Capacity rendering after switching to open) but will test in the wild
for a couple days.

Great reference for new-ish fetch in Nuxt [1] that helped point me in
this alternative direction (see note about mutating local data).

[1]: https://nuxtjs.org/blog/understanding-how-fetch-works-in-nuxt-2-12#new-fetch