Switch back to computed property for occupancy
And set default occupancy state in Vuex store to zero for each key
instead of empty object so initial value for computed property is valid.

This resolves the bug where occupancy data property was not reactive
when the sign changed over from closed to open. So fetch would update
occupancy Vuex store initially, and the Vuex store would be updated
every two minutes via mounted() hook, but data properties were not
reactive so current occupancy was not reflected as it would never change
from initial value returned at opening time.

If you refreshed the page after opening, everything would work as
expected, with occupancy data property fully reactive, but this a deal
breaker for a digital sign that has no user interacting with the
browser.