Ensure Libcal `lid` is set for hours request (#201)
For spaces, when categories are involved, we need to be able to use
either the main hours from the parent library, or override using the
category's hours.

Accidentally introduced bug when implementing building occupancy signs
in #186. As a result, all hours were returned and first in the list
(Africana) was being used for all spaces signs.